Add files via upload
A program that checks if a certain column of a dataframe filled with IPs (string) is in a certain list of IPs and creating a new column in that dataframe indicating T'(True) or 'F'(False). 'T' means that an IP in a row is in the IP list. 'F' means otherwise.  We can use this new 'T'/'F' column to select certain entries while excluding others using GroupBy.